pyproject.toml
[[tool.pdm.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/projects/44844627/packages/pypi/simple"

[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.resolution]
respect-source-order = true

[project]
name = "gwserver"
version = "1.0.0"
description = ""
authors = [{name = "Sergey Pavlov", email = "self@sergeypavlov.dev"}]
requires-python = ">=3.10"
dependencies = [
    "litestar>=2.0.0a3",
    "pydantic>=1.10.7",
    "typer>=0.7.0",
    "uvicorn[standard]>=0.21.1",
    "gunicorn>=20.1.0",
    "alembic>=1.10.3",
    "sqlalchemy>=2.0.9",
    "psycopg2>=2.9.6",
    "boto3>=1.26.108",
    "pandas>=2.0.0",
    "pillow>=9.5.0",
    "onnxruntime-gpu>=1.14.1",
    "numpy>=1.24.2",
    "dramatiq[rabbitmq,watch]>=1.14.2",
    "color-utils>=1.0.1",
]

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.2.0",
    "ruff>=0.0.261",
    "pytest",
]

[project.scripts]
gwserver = "gwserver.cli:entrypoint"

[tool.pdm.scripts]
devserver = "uvicorn gwserver.api:app --reload --host 0.0.0.0 --port 8080"
lint = {composite = ["black gwserver", "isort gwserver", "ruff gwserver", "mypy --strict gwserver"]}

[tool.black]
line-length = 100
quiet = true

[tool.isort]
quiet = true
line_length = 100
multi_line_output = 3
profile = "black"

[tool.ruff]
fix = true
line-length = 100
target-version = "py310"

[tool.mypy]
exclude = "migrations|tasks"
ignore_missing_imports = true
show_error_codes = true
